fixed models class

+++ protocore/src/main/java/com/satispay/protocore/models/generic/LocationV1.java
@@ -2,21 +2,21 @@ package com.satispay.protocore.models.generic;
 
 public class LocationV1 extends Model {
 
-    private double lat;
-    private double lon;
+    private float lat;
+    private float lon;
 
     public LocationV1() { }
 
-    public LocationV1(double lat, double lon) {
+    public LocationV1(float lat, float lon) {
         this.lat = lat;
         this.lon = lon;
     }
 
-    public double getLat() { return lat; }
+    public float getLat() { return lat; }
 
-    public void setLat(double lat) { this.lat = lat; }
+    public void setLat(float lat) { this.lat = lat; }
 
-    public double getLon() { return lon; }
+    public float getLon() { return lon; }
 
-    public void setLon(double lon) { this.lon = lon; }
+    public void setLon(float lon) { this.lon = lon; }
 }
